SASS-8242 Create a separate data type to hold the status
